pyproject.toml
[project]
name = "recruitment-assistant"
version = "0.1.0"
description = "Multi-Agent Recruitment System"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "python-dotenv>=1.2.1",
    "crewai>=0.10.0",
    "fastapi>=0.115.0",
    "streamlit>=1.40.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",
    "pandas>=2.2.0",
    "plotly>=5.24.0",
    "requests>=2.32.0",
    "loguru>=0.7.0",
    "structlog>=24.4.0",
    "fpdf2>=2.8.0",
    "pip>=26.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.7.0",
    "mypy>=1.12.0",
    "pytest>=8.3.0",
    "httpx>=0.28.0",
    "pytest-asyncio>=0.24.0",
    "anyio>=4.6.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
